+++ packages/api-post/src/myPosts.ts
@@ -8,24 +8,20 @@ export interface IMyPostsResponse {
   offset: number
 }
 
+export interface IMyPostsRequest {
+  limit: number
+  offset: number
+  groups?: string[]
+}
+
 export interface IMyPostsDetailsResponse extends IGroup {
   count: number
 }
 
-export const fetchMyPosts = ({
-  limit,
-  offset,
-}: {
-  limit: number
-  offset: number
-}) => {
-  const params = new URLSearchParams()
-  params.set('offset', `${offset}`)
-  params.set('limit', `${limit}`)
-
-  return fetchJSON<void, IMyPostsResponse>(
-    makeApiUrl(`posts/my?${params.toString()}`)
-  )
+export const fetchMyPosts = (payload: IMyPostsRequest) => {
+  return fetchJSON<IMyPostsRequest, IMyPostsResponse>(makeApiUrl('posts/my'), {
+    values: payload,
+  })
 }
 
 export const fetchMyPostDetails = () =>